refactor: controls --> barcontrols, panelcontrols --> controls
